+++ src/astro/astro.hpp
@@ -25,3 +25,5 @@
 #include "delta_t.hpp"
 #include "earth.hpp"
 #include "sun.hpp"
+#include "moon.hpp"
+#include "moon_phase.hpp"