+++ src/astro/julian_day.hpp
@@ -23,7 +23,6 @@
 
 #pragma once
 
-#include <chrono>
 #include <cassert>
 
 #include "delta_t.hpp"