+++ src/test/astro/earth_test.cpp
@@ -73,9 +73,9 @@ TEST(Earth, Vsop87dEvaluate) {
 
   for (const auto& [jde, expected] : dataset) {
     const auto& [λ, β, r] = vsop87d(jde);
-    ASSERT_NEAR(λ.as<DEG>(), std::get<0>(expected), 1e-11);
-    ASSERT_NEAR(β.as<DEG>(), std::get<1>(expected), 1e-16);
-    ASSERT_NEAR(r,           std::get<2>(expected), 1e-14);
+    ASSERT_NEAR(λ.deg(), std::get<0>(expected), 1e-11);
+    ASSERT_NEAR(β.deg(), std::get<1>(expected), 1e-16);
+    ASSERT_NEAR(r.au(),  std::get<2>(expected), 1e-14);
   }
 }